Move some functions to utils.js
